Rename the metric handler source files
