+++ libtenzir/src/metric_handler.cpp
@@ -6,7 +6,7 @@
 // SPDX-FileCopyrightText: (c) 2024 The Tenzir Contributors
 // SPDX-License-Identifier: BSD-3-Clause
 
-#include "tenzir/metrics.hpp"
+#include "tenzir/metric_handler.hpp"
 
 #include <caf/send.hpp>
